+++ extension/src/config/init/init-interface-interaction-custom-type.js
