+++ extension/src/config/init/init-payment-custom-type.js
