+++ extension/src/paymentHandler/creditCard/credit-card.handler.js
@@ -2,9 +2,9 @@ const _ = require('lodash')
 
 const pU = require('../payment-utils')
 
-const creditCardMakePayment = require('./creditCardMakePayment.handler')
-const creditCardCompletePayment = require('./creditCardCompletePayment.handler')
-const errorMessages = require('../../validator/errorMessages')
+const creditCardMakePayment = require('./credit-card-make-payment.handler')
+const creditCardCompletePayment = require('./credit-card-complete-payment.handler')
+const errorMessages = require('../../validator/error-messages')
 
 async function handlePayment (paymentObject) {
   const hasPendingTransaction = _.isObject(pU.getChargeTransactionPending(paymentObject))
