+++ extension/src/paymentHandler/kcp/kcp-make-payment.handler.js
@@ -1,9 +1,9 @@
 const fetch = require('node-fetch')
 
-const c = require('../../config/constants')
 const configLoader = require('../../config/config')
 const pU = require('../payment-utils')
-const ValidatorBuilder = require('../../validator/validatorBuilder')
+const c = require('../../config/constants')
+const ValidatorBuilder = require('../../validator/validator-builder')
 
 const config = configLoader.load()
 
@@ -61,22 +61,23 @@ function _validatePayment (paymentObject) {
 
 async function _callAdyen (paymentObject) {
   const transaction = pU.getChargeTransactionInit(paymentObject)
-  const body = {
-    merchantAccount: config.adyen.merchantAccount,
+  const paymentMethodType = paymentObject.paymentMethodInfo.method
+  const requestBody = {
     amount: {
       currency: transaction.amount.currencyCode,
       value: transaction.amount.centAmount
     },
     reference: paymentObject.interfaceId,
     paymentMethod: {
-      type: 'paypal'
+      type: paymentMethodType
     },
-    returnUrl: paymentObject.custom.fields.returnUrl
+    returnUrl: paymentObject.custom.fields.returnUrl,
+    merchantAccount: config.adyen.merchantAccount
   }
 
   const request = {
     method: 'POST',
-    body: JSON.stringify(body),
+    body: JSON.stringify(requestBody),
     headers: { 'x-api-key': config.adyen.apiKey, 'Content-Type': 'application/json' }
   }
   const response = await fetch(`${config.adyen.apiBaseUrl}/payments`, request)
